Refactored postcode lookup, ordnance survey models, services, and exceptions
