+++ lambdas/postcode-lookup/src/main/java/uk/gov/di/ipv/cri/address/api/exceptions/PostcodeLookupValidationException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.address.library.exception;
+package uk.gov.di.ipv.cri.address.api.exceptions;
 
 public class PostcodeLookupValidationException extends RuntimeException {
     public PostcodeLookupValidationException() {
